[ADD] project with Logs

+++ invoicedoc/src/main/java/com/banquito/core/invoicedoc/service/TaxService.java
@@ -4,6 +4,7 @@ import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
+import java.util.Optional;
 import org.springframework.stereotype.Service;
 
 import com.banquito.core.invoicedoc.dto.TaxDTO;
@@ -11,6 +12,9 @@ import com.banquito.core.invoicedoc.model.Tax;
 import com.banquito.core.invoicedoc.repository.TaxRepository;
 import com.banquito.core.invoicedoc.util.mapper.TaxMapper;
 
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
 @Service
 public class TaxService {
 
@@ -22,10 +26,12 @@ public class TaxService {
         this.taxMapper = taxMapper;
     }
 
-   public TaxDTO createTax(TaxDTO taxDTO) {
+    public TaxDTO createTax(TaxDTO taxDTO) {
+        log.info("Creating tax: {}", taxDTO);
         Tax tax = taxMapper.toPersistence(taxDTO);
         tax.setUniqueId(UUID.randomUUID().toString());
         tax = taxRepository.save(tax);
+        log.info("Tax created: {}", tax);
         return taxMapper.toDTO(tax);
     }
 
@@ -38,25 +44,34 @@ public class TaxService {
     }
 
     public void deleteTax(String id) {
-        taxRepository.deleteById(id);
+        log.info("Deleting tax with ID: {}", id);
+        if (taxRepository.existsById(id)) {
+            taxRepository.deleteById(id);
+            log.info("Tax deleted with ID: {}", id);
+        } else {
+            log.warn("Tax not found with ID: {}", id);
+        }
     }
 
     public TaxDTO getTaxById(String id) {
-        return taxRepository.findById(id)
-                .map(taxMapper::toDTO)
-                .orElse(null);
+        log.info("Fetching tax with ID: {}", id);
+        Optional<Tax> taxOptional = taxRepository.findById(id);
+        if (taxOptional.isPresent()) {
+            log.info("Tax found with ID: {}", id);
+            return taxMapper.toDTO(taxOptional.get());
+        } else {
+            log.warn("Tax not found with ID: {}", id);
+            return null;
+        }
     }
 
     public List<TaxDTO> getAllTaxes() {
-        return taxRepository.findAll().stream()
+        log.info("Fetching all taxes");
+        List<Tax> taxes = taxRepository.findAll();
+        return taxes.stream()
                 .map(taxMapper::toDTO)
                 .collect(Collectors.toList());
     }
 
-    public List<TaxDTO> getTaxesByInvoiceId(String invoiceId) {
-        return taxRepository.findByInvoiceId(invoiceId).stream()
-                .map(taxMapper::toDTO)
-                .collect(Collectors.toList());
-    }
 
 }
